# Use common methods for Base functions instead of generating