--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "Legolas"
 uuid = "741b9549-f6ed-4911-9fbf-4a1c0c97f0cd"
 authors = ["Beacon Biosignals, Inc."]
-version = "0.5.14"
+version = "0.5.15"
 
 [deps]
 Arrow = "69666777-d1a9-59fb-9406-91d4454c9d45"