--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -295,6 +295,13 @@ end
     xs::Union{Vector{String},Missing}
 end
 
+@schema "test.union-missing" UnionMissing
+
+@version UnionMissingV1 begin
+    a::Union{Int,Missing}
+    b::Union{Int,Missing}
+end
+
 @testset "`Legolas.@version` and associated utilities for declared `Legolas.SchemaVersion`s" begin
     @testset "Legolas.SchemaVersionDeclarationError" begin
         @test_throws SchemaVersionDeclarationError("malformed or missing field declaration(s)") eval(:(@version(NewV1, $(Expr(:block, LineNumberNode(1, :test))))))
@@ -517,6 +524,8 @@ end
         @test isequal(c, c)
         @test ismissing(c == c)
         @test hash(c) isa UInt  # NOTE: can't rely on particular values
+        @test UnionMissingV1(; a=missing, b=1) != UnionMissingV1(; a=missing, b=2)
+        @test !isequal(UnionMissingV1(; a=missing, b=1), UnionMissingV1(; a=missing, b=2))
     end
 end