@ericphanson 2023-09-29
Since this closes #101, can you add the test from that post? (The current one doesn't cover the case in which there is `missing` present but `==` should return false since there is a non-missing field that differs)