@jrevels 2023-09-30
can we check that performance is the same as the old methods in benchmark contexts where record types aren't fully inferred