@ararslan 2023-09-30
Not sure if this is what you had in mind but here's what I get. Looks fine to me.

Setup:
```julia
using Legolas, BenchmarkTools
using Legolas: @schema, @version

@schema "test.param" Param
@version ParamV1 begin
    i::(<:Integer)
end
@version ParamV2 begin
    i::(<:Integer)
end

bad1() = ParamV1(; i=(rand() < 0.5 ? Int32 : Int64)(1))
bad2() = (rand() < 0.5 ? ParamV1 : ParamV2)(; i=(rand() < 0.5 ? Int32 : Int64))
```
The return type of `bad1` infers as `Union{ParamV1{Int32},ParamV1{Int64}}` and the return type of `bad2` can't be inferred.

main:
```julia
julia> @benchmark bad1() == bad1()
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  870.300 ns … 37.980 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):       1.292 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.913 μs ±  2.597 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ██▆▆▅▃▂▂▂▂▂▁                                                 ▂
  ███████████████▇▆▅▅▁▄▄▄▄▄▄▄▄▅▅▄▅▄▁▅▁▃▃▁▃▄▁▃▄▁▁▄▁▁▄▃▃▁▃▅▅▃▅▅▅ █
  870 ns        Histogram: log(frequency) by time      19.5 μs <

 Memory estimate: 64 bytes, allocs estimate: 4.

julia> @benchmark bad2() == bad2()
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.441 μs …  14.840 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.660 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.668 μs ± 224.386 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                         ▁▂▂▅▅▆▇▇▇█▇█▇█▆▄▂▃▁▁
  ▁▁▁▁▁▁▁▁▁▂▂▂▂▂▃▃▄▄▅▅▆▆██████████████████████▆▆▅▄▄▃▃▂▂▂▂▂▁▁▁ ▄
  1.44 μs         Histogram: frequency by time        1.83 μs <

 Memory estimate: 96 bytes, allocs estimate: 6.
```

this branch:
```julia
julia> @benchmark bad1() == bad1()
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.269 μs …  11.213 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.493 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.510 μs ± 210.917 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▂▄▆█▇▆▅▂
  ▂▂▂▃▃▅▆▇████████▇▆▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▂▁▁▁▁▁▁▂▂▂▂▂▂▁▂▂▂ ▃
  1.27 μs         Histogram: frequency by time        2.37 μs <

 Memory estimate: 64 bytes, allocs estimate: 4.

julia> @benchmark bad2() == bad2()
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  963.900 ns … 36.246 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):       1.435 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.937 μs ±  2.029 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅█▇▅▅▅▄▃▂▁ ▁▁▁ ▁                                             ▂
  ██████████████████▆█▆▆▆▆▅▄▅▁▁▃▄▄▄▅▄▅▆▅▆▅▆▅▅▃▄▄▄▃▄▅▄▄▅▅▄▄▄▄▃▄ █
  964 ns        Histogram: log(frequency) by time      13.8 μs <

 Memory estimate: 96 bytes, allocs estimate: 6.
```