@ararslan 2023-09-30
More directly targeting the equality check, there's unsurprisingly no difference at all.

main:
```julia
julia> @benchmark x == y setup=(x = bad1(); y = bad1())
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  6.113 ns … 366.950 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.207 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.007 ns ±   8.376 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █  ▆           ▁▅▁   ▃                                      ▁
  █▅▇██▆▆▃▆▅▄▇▇▆▆███▆▅▄██▅▄▄▅▆▅▆▆▆▆▆▆████▇▇█▇▇▆▇▆▆▅▆▅▄▄▄▅▆▇▇▆ █
  6.11 ns      Histogram: log(frequency) by time      17.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark x == y setup=(x = bad2(); y = bad2())
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):  26.546 ns … 97.779 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     34.875 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   35.495 ns ±  4.587 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

               █▃       ▁▆▃
  ▂▃▃▄▃▂▁▁▂▂▂▃▆██▆▅▄▃▃▂▄████▇▅▆▅▆▅▅▅▄▄▄▅▅▆▅▄▄▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▁ ▃
  26.5 ns         Histogram: frequency by time        46.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

this branch:
```julia
julia> @benchmark x == y setup=(x = bad1(); y = bad1())
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  6.113 ns … 192.035 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.820 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.461 ns ±   4.978 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █  █▆▁▁     ▁▁▂▂▂▃▂▂▁ ▁     ▁▁▁                             ▂
  █████████████████████████▇▇██████▇▇▇▆▆▆▆▆▄▆▁▄▆▅▆▄▄▇▄▄▄▄▃▅▅▅ █
  6.11 ns      Histogram: log(frequency) by time      16.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark x == y setup=(x = bad2(); y = bad2())
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):  26.282 ns … 201.913 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     34.388 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   35.156 ns ±   4.858 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                █▅      ▁▅█▂    ▁
  ▂▃▅▃▃▁▁▁▂▄▅▅▄▆██▅▄▅▃▃▅████▇▅▄▅█▇▅▅▆▅▄▅▆▇▆▅▄▄▃▄▃▄▄▃▃▂▂▂▂▂▂▂▂▁ ▃
  26.3 ns         Histogram: frequency by time         45.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```