@jrevels 2023-10-03
I sanity checked NamedTuple conversion performance as it can end up being in the critical inner loops in some use cases; result is that it seems fine on julia 1.9.

```jl
using Legolas, BenchmarkTools
using Legolas: @schema, @version

@schema "test.foo" Foo

@version FooV1 begin
    a::(<:Any)
    b::(<:Any)
    c::(<:Any)
    d::(<:Any)
end

@version FooV2 begin
    x::(<:Any)
    y::(<:Any)
    z::(<:Any)
end

# attempt to thwart compiler optimizations
function test(v, n)
    result = 0
    for _ in 1:n
        result += sum(sizeof, values(NamedTuple(rand(v))))
    end
    return result
end

gen_foos() = Any[FooV1(a=1,b=2,c=3,d=4),FooV2(x="x",y="y",z="z")]

@benchmark test(foos, 100) setup=(foos=gen_foos()) evals=1
```

this branch:

```julia
julia> @benchmark test(foos, 100) setup=(foos=gen_foos()) evals=1
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.541 μs …  1.652 ms  ┊ GC (min … max): 0.00% … 97.53%
 Time  (median):     16.167 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   16.839 μs ± 28.094 μs  ┊ GC (mean ± σ):  2.84% ±  1.69%

        ▄ █▁▆▂
  ▁▁▁▂▃▆█▇████▄▅▂▃▃▂▃▂▃▂▁▂▁▁▁▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  15.5 μs         Histogram: frequency by time        19.5 μs <

 Memory estimate: 15.11 KiB, allocs estimate: 441.
```

main:

```jl
julia> @benchmark test(foos, 100) setup=(foos=gen_foos()) evals=1
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.750 μs …  1.811 ms  ┊ GC (min … max): 0.00% … 96.41%
 Time  (median):     16.625 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.324 μs ± 30.596 μs  ┊ GC (mean ± σ):  2.99% ±  1.68%

          ▄ ██▅
  ▂▂▂▂▂▃▅██████▆█▅▄▄▃▄▄▃▂▂▂▂▂▂▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▂▂▂▂▂▂▂▂▂▂ ▃
  15.8 μs         Histogram: frequency by time        20.1 μs <

 Memory estimate: 14.98 KiB, allocs estimate: 443.
```