@ericphanson 2023-10-03
FWIW if you are looking for perf drops here, I would look at > 10 fields. E.g. in https://discourse.julialang.org/t/slowness-of-fieldnames-and-propertynames/55364/2 (very similar case) I found inference failed for 11 fields but not for 9. That was on an older version of julia though.