@ararslan 2023-10-03
Tried using row types with 15 and 25 parameters and still get no meaningful difference.

setup
```julia
using Legolas, BenchmarkTools, Base.Iterators
using Legolas: @schema, @version

macro make_version(name, n)
    block = Expr(:block)
    ex = Expr(:macrocall, Symbol("@version"), __source__, name, block)
    for i in 1:n
        x = Symbol(:x, i)
        push!(block.args, Expr(:(::), Symbol(:x, i), :(<:Any)))
    end
    return ex
end

@schema "test.foo" Foo
@make_version FooV1 15
@make_version FooV2 25

function make_row(T, values)
    nt = NamedTuple()
    for (field, value) in zip(fieldnames(T), values)
        nt = merge(nt, (; field => value))
    end
    return T(; nt...)
end

gen_foos() = Any[make_row(FooV1, countfrom(1)), make_row(FooV2, string.('a':'z'))]

function test(v, n)
    result = 0
    for _ in 1:n
        result += sum(sizeof, values(NamedTuple(rand(v))))
    end
    return result
end

@benchmark test(foos, 100) setup=(foos = gen_foos()) evals=1
```

this branch
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  17.958 μs … 887.416 μs  ┊ GC (min … max): 0.00% … 95.49%
 Time  (median):     19.166 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.348 μs ±  27.253 μs  ┊ GC (mean ± σ):  4.52% ±  3.29%

     ▁▃▅▅▇███▇▇▅▄▃▂▁                                           ▂
  ▂▆██████████████████▇▇▆▆▅▅▄▆▆▇▆▇▇▇████▇▇▆▇▇▇▇▇▅▆▆▄▆▆▅▃▆▅▅▅▅▄ █
  18 μs         Histogram: log(frequency) by time      25.1 μs <

 Memory estimate: 61.17 KiB, allocs estimate: 484.
```

main
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  19.125 μs … 811.000 μs  ┊ GC (min … max): 0.00% … 94.50%
 Time  (median):     20.292 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.633 μs ±  29.285 μs  ┊ GC (mean ± σ):  4.93% ±  3.55%

         ▁▁▇█▆▂▁
  ▁▁▁▁▂▃▅███████▅▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  19.1 μs         Histogram: frequency by time         25.8 μs <

 Memory estimate: 60.22 KiB, allocs estimate: 485.
```